+++ Source/PathFinder.h
@@ -1,8 +1,6 @@
 #ifndef FINDER_H
 #define FINDER_H
-#include <stdint.h>
 
-
-void convRead(int pocz, int kon, uint8_t ilep0, char *filename, char *zapis);
+void convRead(int pocz, int kon, int ilep0, char *filename, char *zapis);
 
 #endif
